% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{analyte_annotations}
\alias{analyte_annotations}
\title{F- and P-values combined, long format; Data frame containing the data for
the paper}
\format{
A tibble
}
\usage{
analyte_annotations
}
\description{
Object is a tibble (tbl_df)
}
\keyword{datasets}
