% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clean_grinta_zonulin.R
\name{data_clean_grinta_zonulin}
\alias{data_clean_grinta_zonulin}
\alias{zonulin}
\title{loads a clean zonulin dataset for the GRINTA! study}
\usage{
data_clean_grinta_zonulin()
}
\description{
loads a clean zonulin dataset for the GRINTA! study
}
