% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_paper_all_final}
\alias{data_paper_all_final}
\title{All summarised data from the paper, long format; Data frame containing the data for
the paper}
\format{
A tibble
}
\usage{
data_paper_all_final
}
\description{
Object is a tibble (tbl_df)
}
\keyword{datasets}
