% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{f_values_long_protocol}
\alias{f_values_long_protocol}
\title{F-values; Data frame containing the data for the paper}
\format{
A tibble
}
\usage{
f_values_long_protocol
}
\description{
Object is a tibble (tbl_df)
}
\keyword{datasets}
