% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_qq.R
\name{gg_qq}
\alias{gg_qq}
\title{ggplot implementation of QQ normality plot}
\usage{
gg_qq(
  x,
  distribution = "norm",
  ...,
  line.estimate = NULL,
  conf = 0.95,
  labels = names(x)
)
}
\description{
ggplot implementation of QQ normality plot
}
