% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_values_long_protocol}
\alias{p_values_long_protocol}
\title{P-values; Data frame containing the data for the paper}
\format{
A tibble
}
\usage{
p_values_long_protocol
}
\description{
Object is a tibble (tbl_df)
}
\keyword{datasets}
