% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panel_plot_list}
\alias{panel_plot_list}
\title{A list of ggplot2 objects for displaying graphs in panel plots}
\format{
List
}
\usage{
panel_plot_list
}
\description{
A list of ggplot2 objects for displaying graphs in panel plots
}
\keyword{graphs}
