% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rpe_scores}
\alias{rpe_scores}
\title{RPE scores; Data frame containing the data for the paper}
\format{
A tibble
}
\usage{
rpe_scores
}
\description{
Object is a tibble (tbl_df)
}
\keyword{datasets}
